Move test script to its own pytest test without much modification so far
